refactor: Update db seeds; add user accounts for unit tests
