+++ .vscode/db.code-snippets
@@ -3,7 +3,7 @@
     "scope": "javascript,typescript",
     "prefix": "migration",
     "body": [
-      "import { type Knex } from \"knex\";",
+      "import { Knex } from \"knex\";",
       "",
       "/**",
       " * Migrates database schema to the next version.",
