--- db/seeds/02_identities.json
@@ -1,1 +1,0 @@
-[]
