+++ db/seeds/identities.json
@@ -1,0 +1,13 @@
+[
+  {
+    "provider": "google",
+    "id": "108783264480242316545",
+    "user_id": "wp60xu",
+    "username": null,
+    "email": "koistya@tarkus.me",
+    "profile": {},
+    "credentials": {},
+    "created": "2019-01-01T00:00:00.000Z",
+    "updated": "2019-01-01T00:00:00.000Z"
+  }
+]
